--- a/src/get_SOA_table.jl
+++ b/src/get_SOA_table.jl
@@ -7,18 +7,18 @@ Given the id or name of a `mort.SOA.org` table, grab it and return it as a `Mort
 !!! Remember that not all tables have been tested to work.
 """
 function get_SOA_table(id::Int)
-    readXTbML(joinpath(artifact"mort.soa.org","t$id.xml"))
+    readXTbML(joinpath(artifact"mort.soa.org", "t$id.xml"))
 end
 
-function get_SOA_table(table_name::String;source_map=table_source_map)
-    entry = findfirst(x-> x.name == table_name, source_map)
+function get_SOA_table(table_name::String; source_map = table_source_map)
+    entry = get(source_map, table_name, nothing)
     if entry === nothing
         search_method = StringDistances.Partial(StringDistances.Levenshtein())
-        suggestion,_ = StringDistances.findnearest(table_name,[x.name for x in table_source_map],search_method)
+        suggestion, _ = StringDistances.findnearest(table_name, [x.name for x in table_source_map], search_method)
         throw(ArgumentError("table name \"table_name\" not found in table set; " *
-            "most similar available name is: \"$suggestion\""))
+                            raw"most similar available name is: \"$suggestion\""))
     end
-    readXTbML(joinpath(artifact"mort.soa.org","t$(source_map[entry].id).xml"))
+    readXTbML(joinpath(artifact"mort.soa.org", "t$entry.xml"))
 end
 
 """