Memoize table build so that subsequent calls are much faster:

Before: 
```
julia> @benchmark MortalityTables.table("1980 CSO Basic Table – Male, ANB")
BenchmarkTools.Trial: 10000 samples with 1 evaluation.
 Range (min … max):  256.458 μs …  20.328 ms  ┊ GC (min … max): 0.00% … 42.11%
 Time  (median):     268.041 μs               ┊ GC (median):    0.00%
 Time  (mean ± σ):   339.263 μs ± 982.478 μs  ┊ GC (mean ± σ):  7.75% ±  2.62%

      ▆█▇▄▂                                                      
  ▁▂▃███████▆▆▆▅▄▃▃▂▂▂▂▂▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁ ▂
  256 μs           Histogram: frequency by time          350 μs <

 Memory estimate: 213.61 KiB, allocs estimate: 3441.
```

after:
```
 julia> @benchmark MortalityTables.table("1980 CSO Basic Table – Male, ANB")
BenchmarkTools.Trial: 10000 samples with 5 evaluations.
 Range (min … max):  6.358 μs … 478.325 μs  ┊ GC (min … max): 0.00% … 97.82%
 Time  (median):     6.608 μs               ┊ GC (median):    0.00%
 Time  (mean ± σ):   7.047 μs ±  12.119 μs  ┊ GC (mean ± σ):  4.78% ±  2.75%

      ▆█▇▇▄▂                                                   
  ▁▂▄███████▇▆▄▄▄▃▃▃▂▂▂▂▂▂▂▁▂▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁ ▂
  6.36 μs         Histogram: frequency by time         8.3 μs <

 Memory estimate: 7.28 KiB, allocs estimate: 64.
```